+++ crates/feos/benches/dual_numbers.rs
@@ -13,7 +13,6 @@ use feos_core::parameter::PureRecord;
 use nalgebra::{DVector, Dyn, dvector};
 use num_dual::{Dual2_64, Dual3_64, Dual64, DualNum, HyperDual64};
 use quantity::*;
-use typenum::P3;
 
 /// Helper function to create a state for given parameters.
 /// - temperature is 80% of critical temperature,
@@ -129,7 +128,7 @@ fn methane_co2_pcsaft(c: &mut Criterion) {
 
     // 230 K, 50 bar, x0 = 0.15
     let temperature = 230.0 * KELVIN;
-    let density = 24.16896 * KILO * MOL / METER.powi::<P3>();
+    let density = 24.16896 * KILO * MOL / METER.powi::<3>();
     let volume = 10.0 * MOL / density;
     let x = dvector![0.15, 0.85];
     let moles = &x * 10.0 * MOL;