+++ crates/feos/tests/pcsaft/critical_point.rs
@@ -6,7 +6,6 @@ use nalgebra::dvector;
 use quantity::*;
 use std::error::Error;
 use std::sync::Arc;
-use typenum::P3;
 
 #[test]
 fn test_critical_point_pure() -> Result<(), Box<dyn Error>> {
@@ -22,7 +21,7 @@ fn test_critical_point_pure() -> Result<(), Box<dyn Error>> {
     assert_relative_eq!(cp.temperature, 375.12441 * KELVIN, max_relative = 1e-8);
     assert_relative_eq!(
         cp.density,
-        4733.00377 * MOL / METER.powi::<P3>(),
+        4733.00377 * MOL / METER.powi::<3>(),
         max_relative = 1e-6
     );
     Ok(())
@@ -43,7 +42,7 @@ fn test_critical_point_mix() -> Result<(), Box<dyn Error>> {
     assert_relative_eq!(cp.temperature, 407.93481 * KELVIN, max_relative = 1e-8);
     assert_relative_eq!(
         cp.density,
-        4265.50745 * MOL / METER.powi::<P3>(),
+        4265.50745 * MOL / METER.powi::<3>(),
         max_relative = 1e-6
     );
     Ok(())